+++ Tests/Polynomials/ModintPolynomial/TaylorShift.test.cpp
@@ -11,7 +11,7 @@ using namespace std;
 int main() {
     ios_base::sync_with_stdio(false);
     cin.tie(nullptr);
-    int n, c;
+    long long n, c;
     cin >> n >> c;
     vector<mint> a(n);
     for (auto& x: a)
